@@ -27,12 +27,14 @@
     ],
     "require": {
         "ext-openssl": "*",
-        "php": "^8.0",
-        "payum/core": "^1.5",
-        "symfony/options-resolver": "^7.2"
+        "ext-json": "*",
+        "php": "^8.1",
+        "payum/core": "^1.7",
+        "php-http/message-factory": "^1.1"
     },
     "require-dev": {
-        "php-http/guzzle6-adapter": "^1.0"
+        "php-http/guzzle6-adapter": "^1.0",
+        "phpunit/phpunit": "^10"
     },
     "config": {
         "bin-dir": "bin"
@@ -42,9 +44,14 @@
             "MrcMorales\\Payum\\Redsys\\": "src"
         }
     },
+    "autoload-dev": {
+        "psr-4": {
+            "MrcMorales\\Payum\\Redsys\\Tests\\": "tests"
+        }
+    },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.6-dev"
+            "dev-1.7": "1.7-dev"
         }
     }
 }
