@@ -1,13 +1,13 @@
 <?php
+
 namespace MrcMorales\Payum\Redsys\Action\Api;
 
-
+use MrcMorales\Payum\Redsys\Api;
 use Payum\Core\Action\ActionInterface;
 use Payum\Core\ApiAwareInterface;
 use Payum\Core\ApiAwareTrait;
 use Payum\Core\GatewayAwareInterface;
 use Payum\Core\GatewayAwareTrait;
-use MrcMorales\Payum\Redsys\Api;
 
 abstract class BaseApiAwareAction implements ActionInterface, GatewayAwareInterface, ApiAwareInterface
 {
