@@ -1,23 +1,22 @@
 <?php
+
 namespace MrcMorales\Payum\Redsys\Action;
 
 use Payum\Core\Action\ActionInterface;
 use Payum\Core\Bridge\Spl\ArrayObject;
+use Payum\Core\Exception\RequestNotSupportedException;
 use Payum\Core\GatewayAwareInterface;
 use Payum\Core\GatewayAwareTrait;
 use Payum\Core\Request\Authorize;
-use Payum\Core\Exception\RequestNotSupportedException;
 
 class AuthorizeAction implements ActionInterface, GatewayAwareInterface
 {
     use GatewayAwareTrait;
 
     /**
-     * {@inheritDoc}
-     *
      * @param Authorize $request
      */
-    public function execute($request)
+    public function execute($request): void
     {
         RequestNotSupportedException::assertSupports($this, $request);
 
@@ -26,14 +25,11 @@ class AuthorizeAction implements ActionInterface, GatewayAwareInterface
         throw new \LogicException('Not implemented');
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function supports($request)
+    public function supports($request): bool
     {
         return
-            $request instanceof Authorize &&
-            $request->getModel() instanceof \ArrayAccess
+            $request instanceof Authorize
+            && $request->getModel() instanceof \ArrayAccess
         ;
     }
 }
