@@ -3,56 +3,62 @@
 namespace MrcMorales\Payum\Redsys\Action;
 
 use MrcMorales\Payum\Redsys\Action\Api\BaseApiAwareAction;
+use MrcMorales\Payum\Redsys\Api;
+use MrcMorales\Payum\Redsys\Util\TransactionType;
 use Payum\Core\Action\ActionInterface;
 use Payum\Core\Bridge\Spl\ArrayObject;
+use Payum\Core\Exception\RequestNotSupportedException;
 use Payum\Core\GatewayAwareTrait;
 use Payum\Core\Reply\HttpPostRedirect;
 use Payum\Core\Request\Capture;
-use Payum\Core\Exception\RequestNotSupportedException;
-use MrcMorales\Payum\Redsys\Api;
+use Payum\Core\Security\GenericTokenFactoryAwareInterface;
+use Payum\Core\Security\GenericTokenFactoryAwareTrait;
 
-class CaptureAction extends BaseApiAwareAction implements ActionInterface
+class CaptureAction extends BaseApiAwareAction implements ActionInterface, GenericTokenFactoryAwareInterface
 {
     use GatewayAwareTrait;
+    use GenericTokenFactoryAwareTrait;
 
     /** @var Api */
     protected $api;
 
-
     /**
-     * {@inheritDoc}
-     *
      * @param Capture $request
      */
-    public function execute($request)
+    public function execute($request): void
     {
         RequestNotSupportedException::assertSupports($this, $request);
 
-        /** @var  $postData */
         $postData = ArrayObject::ensureArrayObject($request->getModel());
 
         if (empty($postData['Ds_Merchant_MerchantURL']) && $request->getToken()) {
-            $postData['Ds_Merchant_MerchantURL'] = $request->getToken()->getTargetUrl();
+            $notifyToken = $this->tokenFactory->createNotifyToken(
+                $request->getToken()->getGatewayName(),
+                $request->getToken()->getDetails()
+            );
+            $postData['Ds_Merchant_MerchantURL'] = $notifyToken->getTargetUrl();
         }
 
-        $postData->validatedKeysSet(array(
+        $postData->validatedKeysSet([
             'Ds_Merchant_Amount',
             'Ds_Merchant_Order',
             'Ds_Merchant_Currency',
             'Ds_Merchant_TransactionType',
             'Ds_Merchant_MerchantURL',
-        ));
+        ]);
 
-        if (false === $postData['Ds_Merchant_UrlOK'] && $request->getToken()) {
+        $postData['Ds_Merchant_TransactionType'] = TransactionType::AUTHORIZATION;
+
+        if (!$postData['Ds_Merchant_UrlOK'] && $request->getToken()) {
             $postData['Ds_Merchant_UrlOK'] = $request->getToken()
-                ->getTargetUrl();
-        }
-        if (false === $postData['Ds_Merchant_UrlKO'] && $request->getToken()) {
-            $postData['Ds_Merchant_UrlKO'] = $request->getToken()
-                ->getTargetUrl();
+                ->getAfterUrl();
         }
 
-        $details['Ds_Merchant_TransactionType'] = $this->api::TRANSACTIONTYPE_AUTHORIZATION;
+        if (!$postData['Ds_Merchant_UrlKO'] && $request->getToken()) {
+            $postData['Ds_Merchant_UrlKO'] = $request->getToken()
+                ->getAfterUrl();
+        }
+
         $details['Ds_SignatureVersion'] = Api::SIGNATURE_VERSION;
         $details['Ds_MerchantParameters'] = $this->api->createMerchantParameters($postData->toUnsafeArray());
         $details['Ds_Signature'] = $this->api->sign($postData->toUnsafeArray());
@@ -60,15 +66,10 @@ class CaptureAction extends BaseApiAwareAction implements ActionInterface
         throw new HttpPostRedirect($this->api->getApiEndpoint(), $details);
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function supports($request)
+    public function supports($request): bool
     {
         return
-            $request instanceof Capture &&
-            $request->getModel() instanceof \ArrayAccess;
+            $request instanceof Capture
+            && $request->getModel() instanceof \ArrayAccess;
     }
 }
-
-
