@@ -1,8 +1,10 @@
 <?php
+
 namespace MrcMorales\Payum\Redsys\Action;
 
 use MrcMorales\Payum\Redsys\Action\Api\BaseApiAwareAction;
 use MrcMorales\Payum\Redsys\Api;
+use MrcMorales\Payum\Redsys\Util\TransactionType;
 use Payum\Core\Action\ActionInterface;
 use Payum\Core\Bridge\Spl\ArrayObject;
 use Payum\Core\Exception\RequestNotSupportedException;
@@ -18,11 +20,9 @@ class ConvertPaymentAction extends BaseApiAwareAction implements ActionInterface
     protected $api;
 
     /**
-     * {@inheritDoc}
-     *
      * @param Convert $request
      */
-    public function execute($request)
+    public function execute($request): void
     {
         RequestNotSupportedException::assertSupports($this, $request);
 
@@ -31,28 +31,24 @@ class ConvertPaymentAction extends BaseApiAwareAction implements ActionInterface
 
         $details = ArrayObject::ensureArrayObject($payment->getDetails());
 
-        $details->defaults(array(
+        $details->defaults([
             'Ds_Merchant_Amount' => $payment->getTotalAmount(),
             'Ds_Merchant_Order' => $this->api->ensureCorrectOrderNumber($payment->getNumber()),
             'Ds_Merchant_MerchantCode' => $this->api->getMerchantCode(),
             'Ds_Merchant_Currency' => $this->api->getISO4127($payment->getCurrencyCode()),
             'Ds_Merchant_Terminal' => $this->api->getMerchantTerminalCode(),
-            'Ds_Merchant_TransactionType' => $this->api::TRANSACTIONTYPE_AUTHORIZATION,
-        ));
+            'Ds_Merchant_TransactionType' => TransactionType::AUTHORIZATION,
+        ]);
 
-
-        $request->setResult((array)$details);
+        $request->setResult((array) $details);
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function supports($request)
+    public function supports($request): bool
     {
         return
-            $request instanceof Convert &&
-            $request->getSource() instanceof PaymentInterface &&
-            $request->getTo() == 'array'
+            $request instanceof Convert
+            && $request->getSource() instanceof PaymentInterface
+            && 'array' == $request->getTo()
         ;
     }
 }
