@@ -1,23 +1,21 @@
 <?php
+
 namespace MrcMorales\Payum\Redsys\Action;
 
+use MrcMorales\Payum\Redsys\Api;
 use Payum\Core\Action\ActionInterface;
-use Payum\Core\Request\GetStatusInterface;
 use Payum\Core\Bridge\Spl\ArrayObject;
 use Payum\Core\Exception\RequestNotSupportedException;
-use MrcMorales\Payum\Redsys\Api;
+use Payum\Core\Request\GetStatusInterface;
 
 class StatusAction implements ActionInterface
 {
     /**
-     * {@inheritDoc}
-     *
      * @param GetStatusInterface $request
      */
-    public function execute($request)
+    public function execute($request): void
     {
         RequestNotSupportedException::assertSupports($this, $request);
-
         $model = ArrayObject::ensureArrayObject($request->getModel());
 
         if (null == $model['Ds_Response']) {
@@ -26,7 +24,6 @@ class StatusAction implements ActionInterface
             return;
         }
 
-
         if ($model['Ds_AuthorisationCode'] && null === $model['Ds_Response']) {
             $request->markPending();
 
@@ -34,7 +31,7 @@ class StatusAction implements ActionInterface
         }
 
         if (in_array($model['Ds_Response'],
-            array(Api::DS_RESPONSE_CANCELED, Api::DS_RESPONSE_USER_CANCELED))) {
+            [Api::DS_RESPONSE_CANCELED, Api::DS_RESPONSE_USER_CANCELED])) {
             $request->markCanceled();
 
             return;
@@ -47,18 +44,13 @@ class StatusAction implements ActionInterface
         }
 
         $request->markUnknown();
-
-
     }
 
-    /**
-     * {@inheritDoc}
-     */
-    public function supports($request)
+    public function supports($request): bool
     {
         return
-            $request instanceof GetStatusInterface &&
-            $request->getModel() instanceof \ArrayAccess
+            $request instanceof GetStatusInterface
+            && $request->getModel() instanceof \ArrayAccess
         ;
     }
 }
